+++ facetorch/analyzer/reader/__init__.py
@@ -1,3 +1,3 @@
-from .core import ImageReader, TensorReader
+from .core import ImageReader, TensorReader, UniversalReader
 
-__all__ = ["ImageReader", "TensorReader"]
+__all__ = ["ImageReader", "TensorReader", "UniversalReader"]