+++ tests/test_au.py
@@ -16,6 +16,5 @@ def test_lip_pucker(response, cfg):
     if "test.jpg" not in cfg.path_image:
         pytest.skip("Only test.jpg is used for this test.")
     if hasattr(cfg, "path_tensor"):
-        if "tensor.pt" in cfg.path_tensor:
-            pytest.skip("Only test.jpg is used for this test.")
+        pytest.skip("Only test.jpg is used for this test.")
     assert response.faces[1].preds["au"].label == "lip_pucker"